+++ tsdef/index.d.ts
@@ -80,6 +80,7 @@ export {
   StatsReport,
   TrackStats,
   TrackSwitchOffMode,
+  TwilioWarning,
   VideoBandwidthProfileOptions,
   VideoCodec,
   VideoCodecSettings,