+++ tsdef/TwilioWarning.d.ts
@@ -1,0 +1,3 @@
+export enum TwilioWarning {
+  recordingMediaLost = 'recording-media-lost'
+}