+++ tsdef/index.d.ts
@@ -42,6 +42,7 @@ export { Room } from './Room';
 export { Track } from './Track';
 export { TrackPublication } from './TrackPublication';
 export { TwilioError } from './TwilioError';
+export { TwilioWarning } from './TwilioWarning';
 export {
   AudioCodec,
   AudioCodecSettings,