+++ tsdef/types.d.ts
@@ -12,6 +12,7 @@ import { RemoteDataTrackPublication } from './RemoteDataTrackPublication';
 import { RemoteVideoTrack } from './RemoteVideoTrack';
 import { RemoteVideoTrackPublication } from './RemoteVideoTrackPublication';
 import { Track } from './Track';
+import { TwilioWarning } from './TwilioWarning';
 import { VideoTrack } from './VideoTrack';
 
 export type LocalTrack = LocalAudioTrack | LocalVideoTrack | LocalDataTrack;
@@ -188,6 +189,7 @@ export interface ConnectOptions {
   maxVideoBitrate?: number | null;
   name?: string | null;
   networkQuality?: boolean | NetworkQualityConfiguration;
+  notifyWarnings?: Array<TwilioWarning>;
   region?: string;
   preferredAudioCodecs?: Array<AudioCodec | AudioCodecSettings | OpusCodecSettings>;
   preferredVideoCodecs?: Array<VideoCodec | VideoCodecSettings | VP8CodecSettings> | VideoEncodingMode;