+++ CHANGELOG.md
@@ -2,6 +2,66 @@ The Twilio Programmable Video SDKs use [Semantic Versioning](http://www.semver.o
 
 **Version 1.x reached End of Life on September 8th, 2021.** See the changelog entry [here](https://www.twilio.com/changelog/end-of-life-complete-for-unsupported-versions-of-the-programmable-video-sdk). Support for the 1.x version ended on December 4th, 2020.
 
+2.22.0 (In Progress)
+====================
+
+New Features
+------------
+
+This release include the **Media Warnings API (Beta)** to help surface media related warning events on the SDK whenever the media server is not able to detect media from a published audio or video track.
+
+### Example
+
+```js
+const room = await connect('token', {
+  notifyWarnings: [ 'recording-media-lost' ]
+  // Other connect options
+});
+
+Array.from(room.localParticipant.tracks.values()).forEach(publication => {
+  publication.on('warning', name => {
+    if (name === 'recording-media-lost') {
+      console.log(`LocalTrack ${publication.track.name} is not recording media.`);
+
+      // Wait a reasonable amount of time to clear the warning.
+      const timer = setTimeout(() => {
+        // If the warning is not cleared, you can manually
+        // reconnect to the room, or show a dialog to the user
+      }, 5000);
+
+      publication.once('warningsCleared', () => {
+        console.log(`LocalTrack ${publication.track.name} warnings have cleared!`);
+        clearTimeout(timer);
+      });
+    }
+  });
+});
+```
+
+### API Definitions
+
+#### ConnectOptions
+
+- **notifyWarnings** - An array of TwilioWarnings to listen to. By default, this array is empty and no warning events will be raised. Possible TwilioWarning values include:
+
+  - `recording-media-lost` - Raised when the media server has not detected any media on the published track that is being recorded in the past 30 seconds.
+
+#### Events
+
+The SDK raises warning events when it detects certain conditions (defined in `TwilioWarning` enumeration). You can implement callbacks on these events to act on them, or to alert the user of an issue. Subsequently, "warningsCleared" event is raised when conditions have returned to normal.
+
+- **LocalTrackPublication.on('warning', callback(name))** - Raised when the published Track encounters a TwilioWarning.
+
+- **LocalTrackPublication.on('warningsCleared', callback())** - Raised when the published Track cleared all TwilioWarning.
+
+- **LocalParticipant.on('trackWarning', callback(name, publication))** - Raised when one of the LocalParticipant's published tracks encounters a TwilioWarning.
+
+- **LocalParticipant.on('trackWarningsCleared', callback(publication))** - Raised when one of the LocalParticipant's published tracks cleared all TwilioWarnings.
+
+- **Room.on('trackWarning', callback(name, publication, participant))** - Raised when one of the LocalParticipant's published tracks in the Room encounters a TwilioWarning.
+
+- **Room.on('trackWarningsCleared', callback(publication, participant))** - Raised when one of the LocalParticipant's published tracks in the Room cleared all TwilioWarnings.
+
 2.21.0 (March 8, 2022)
 ==========================